*ScatterDraw: Fixed Null error
git-svn-id: svn://ultimatepp.org/upp/trunk@13574 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/ScatterDraw/DataSource.h
@@ -388,7 +388,7 @@ private:
 	int idx = 0, idy = 1;
 	Vector<int> idsx, idsy, idsFixed;
 	int beginData = 0;
-	int64 numData = Null;
+	int numData = Null;
 	
 public:
 	VectorVectorY() {}
